binproject/requirements.txt
djangorestframework
python-dotenv
psycopg
# psycopg2-binary
djangorestframework-simplejwt